A new structure for wagtail.wagtailforms.tests
